# openapi/beecount-cloud-v1.yaml
openapi: 3.1.0
info:
  title: BeeCount Cloud
  description: BeeCount Cloud v1 API
  version: 0.1.0
paths:
...
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogoutRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Logout Api V1 Auth Logout Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/devices:
    get:
      tags:
      - devices
      summary: List Devices
      operationId: list_devices_api_v1_devices_get
      parameters:
      - name: view
        in: query
        required: false
        schema:
          enum:
          - deduped
          - sessions
          type: string
          default: deduped
          title: View
      - name: active_within_days
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 30
          title: Active Within Days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DeviceOut'
                type: array
                title: Response List Devices Api V1 Devices Get
      security:
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - app_write
  /api/v1/devices/{device_id}/revoke:
    post:
      tags:
      - devices
      summary: Revoke Device
      operationId: revoke_device_api_v1_devices__device_id__revoke_post
      security:
      - OAuth2PasswordBearer:
        - ops_write
      - OAuth2PasswordBearer:
        - app_write
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Revoke Device Api V1 Devices  Device Id  Revoke Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/devices:
    get:
      tags:
      - admin
      summary: List Devices
      operationId: list_devices_api_v1_admin_devices_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: online_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Online Only
      - name: active_within_days
        in: query
        required: false
        schema:
          type: integer
          maximum: 3650
          minimum: 0
          default: 30
          title: Active Within Days
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDeviceListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AdminDeviceListOut:
      properties:
        total:
          type: integer
          title: Total
        items:
          items:
            $ref: '#/components/schemas/AdminDeviceOut'
          type: array
          title: Items
      type: object
      required:
      - total
      - items
      title: AdminDeviceListOut
    AdminDeviceOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        platform:
          type: string
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        last_ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Ip
        created_at:
          type: string
          format: date-time
          title: Created At
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
        is_online:
          type: boolean
          title: Is Online
        user_id:
          type: string
          title: User Id
        user_email:
          type: string
          title: User Email
      type: object
      required:
      - id
      - name
      - platform
      - created_at
      - last_seen_at
      - is_online
      - user_id
      - user_email
      title: AdminDeviceOut
...
    AuthLoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        platform:
          anyOf:
          - type: string
          - type: 'null'
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        client_type:
          type: string
          enum:
          - app
          - web
          title: Client Type
          default: app
      type: object
      required:
      - email
      - password
      title: AuthLoginRequest
    AuthLogoutRequest:
      properties:
        refresh_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      type: object
      title: AuthLogoutRequest
    AuthRefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: AuthRefreshRequest
    AuthRegisterRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          minLength: 6
          title: Password
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
        device_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Name
        platform:
          anyOf:
          - type: string
          - type: 'null'
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        client_type:
          type: string
          enum:
          - app
          - web
          title: Client Type
          default: app
      type: object
      required:
      - email
      - password
      title: AuthRegisterRequest
    AuthTokenResponse:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        expires_in:
          type: integer
          title: Expires In
        device_id:
          type: string
          title: Device Id
        scopes:
          items:
            type: string
          type: array
          title: Scopes
      type: object
      required:
      - user
      - access_token
      - refresh_token
      - expires_in
      - device_id
      title: AuthTokenResponse
...
    DeviceOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        platform:
          type: string
          title: Platform
        app_version:
          anyOf:
          - type: string
          - type: 'null'
          title: App Version
        os_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Os Version
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        last_ip:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Ip
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
        created_at:
          type: string
          format: date-time
          title: Created At
        session_count:
          type: integer
          title: Session Count
          default: 1
      type: object
      required:
      - id
      - name
      - platform
      - last_seen_at
      - created_at
      title: DeviceOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          title: Email
        is_admin:
          type: boolean
          title: Is Admin
          default: false
      type: object
      required:
      - id
      - email
      title: UserOut
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes:
            app_write: App client read/write scope
            web_read: Web console read scope
            web_write: Web console write scope
            ops_write: Operations and admin scope
          tokenUrl: /api/v1/auth/login